requirements.txt
streamlit==1.45.1
ultralytics==8.3.155
opencv-python-headless==4.9.0.80
Pillow==10.3.0
